+++ src/index.ts
@@ -1,5 +1,4 @@
 export { default as createAction } from './actions/create-action';
 export { default as createReducer } from './reducers/create-reducer';
 export { default as middleware } from './middleware/async-generator-middleware';
-export { failure } from './actions/failure';
 export { nothing } from 'immer';